engine: add direct execution to the process queue.
If an operation was sent to a slave engine or into the own
engine, it was queued, instead of being directly executed.

This causes loss of cycles in case a script or other source
wishes to send operations in the same cycle.

This patch executes the send operation immediately if the
current thread also executes the engine.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>